refactor(detector-aws): add EKS detector implementing DetectorSync interface
